+++ activesupport/test/deprecation_test.rb
@@ -935,6 +935,18 @@ class DeprecationTest < ActiveSupport::TestCase
     end
   end
 
+  class_eval(<<~RUBY, "/path/to/template.html.erb", 1)
+    def generated_method_that_call_deprecation(deprecator)
+      deprecator.warn("Here", caller_locations(0, 10))
+    end
+  RUBY
+
+  test "warn deprecation can blame code generated with eval" do
+    @deprecator.behavior = ->(message, *) { @message = message }
+    generated_method_that_call_deprecation(@deprecator)
+    assert_equal "DEPRECATION WARNING: Here (called from generated_method_that_call_deprecation at /path/to/template.html.erb:2)", @message
+  end
+
   private
     def deprecator_with_messages
       klass = Class.new(ActiveSupport::Deprecation)